add react query provider
